맥주 item ui 수정

+++ app/src/main/java/com/landvibe/beereverything/data/Beer.kt
@@ -11,7 +11,7 @@ data class Beer(
     @ColumnInfo(name = "name") var name: String? = "",
     @ColumnInfo(name = "name_en") var name_en: String? = "",
     @ColumnInfo(name = "country") var country: String? = "",
-    @ColumnInfo(name = "volume") var volume: Int? = 0,
+    @ColumnInfo(name = "volume") var volume: Double? = 0.0,
     @ColumnInfo(name = "category") var category: String? = "",
     @ColumnInfo(name = "favorite") var favorite : Boolean? = false,
     @ColumnInfo(name = "image_url") var image_url : String = ""
